requirements.txt
pinecone_client
python-dotenv
jupyter
pandas
numpy
datasets
uvicorn
gunicorn
fastapi
